Api42/requirements.txt
django==4.2.10
psycopg[binary]
djangorestframework==3.14.0
requests
uvicorn
serializers